Create settings.py
the only change you would be making to the settings.py is giving the path where u store the templates, i.e. TEMPLATE_DIRS, since all the webpages are extended from one template so for any of the webpage to work, the path must be given in TEMPLATE_DIRS in settings file.